feat(task): merge Task and TimedTask and ensure it working with void
