+++ test/lib/executor.cpp
@@ -1,7 +1,7 @@
 #include <catch2/catch.hpp>
 
 #include "../../lib/Executor/Executor.h"
-#include "../../lib/Task/TimedTask.h"
+#include "../../lib/Task/Task.h"
 
 #include <iostream>
 #include <vector>
@@ -10,13 +10,13 @@ using namespace std;
 
 SCENARIO ("Executor basic I/O") {
     GIVEN ("A String executor and a list of strings") {
-        Executor <TimedTask> executor;
+        Executor <Task <void>> executor;
 
         REQUIRE (executor.empty());
         WHEN ("I insert the strings into the executor") {
-            CHECK_NOTHROW (executor.schedule (std::make_unique <TimedTask> ([] { LOG (INFO) << "task 1/3 successful"; })));
-            CHECK_NOTHROW (executor.schedule (std::make_unique <TimedTask> ([] { LOG (INFO) << "task 2/3 successful"; })));
-            CHECK_NOTHROW (executor.schedule (std::make_unique <TimedTask> ([] { LOG (INFO) << "task 3/3 successful"; })));
+            CHECK_NOTHROW (executor.schedule (Task <void> ([] { LOG (INFO) << "task 1/3 successful"; })));
+            CHECK_NOTHROW (executor.schedule (Task <void> ([] { LOG (INFO) << "task 2/3 successful"; })));
+            CHECK_NOTHROW (executor.schedule (Task <void> ([] { LOG (INFO) << "task 3/3 successful"; })));
         }
         CHECK_NOTHROW (executor.await());
         REQUIRE (executor.empty());
